Return bristindex as float

+++ app/Aggregators/Yrkesstatistik/YrkesgruppAggregator.php
@@ -106,6 +106,6 @@ class YrkesgruppAggregator extends BaseAggregator
 
     public static function round($value)
     {
-        return number_format(round($value, 2), 2, '.', '');
+        return (float) number_format(round($value, 2), 2, '.', '');
     }
 }
